requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2
pillow
python-multipart
pika
httpx
